refactor: Domain 계층이 presentation 계층을 의존하는 문제 해결
