+++ src/main/java/org/project/ssogssog/domain/stockmetric/enums/MetricBasePeriod.java
@@ -1,0 +1,6 @@
+package org.project.ssogssog.domain.stockmetric.enums;
+
+public enum MetricBasePeriod {
+    PREV_YEAR,
+    PREV_QUARTER,
+}
