--- src/main/java/org/project/ssogssog/presentation/controller/stockmetric/enums/MetricBasePeriod.java
@@ -1,6 +1,0 @@
-package org.project.ssogssog.presentation.controller.stockmetric.enums;
-
-public enum MetricBasePeriod {
-    PREV_YEAR,
-    PREV_QUARTER,
-}
